Remove manual creation of tmp.mount.
The tmpfs handler will run systemctl enable tmp.mount, which will create a link to this unit file. So we don't need to create it.
If we do, then the handler fails when trying to create the link.
We do need a place to notify the handler to run, so leave the directory task there.